Backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
azure-storage-blob
